+++ sdk/storage/azure-storage-blob-batch/CHANGELOG.md
@@ -7,7 +7,7 @@
 ### Breaking Changes
 
 ### Bugs Fixed
-- Fixed an issue where `BlobBatchSetBlobAccessTierOptions` would not properly handle encoded blob names.
+- Fixed an issue where `BlobBatchSetBlobAccessTierOptions` would not properly handle blob names with special characters.
 
 ### Other Changes
 - Added support for container names with special characters when using OneLake.